Create a new submodule for functions on operations on the Config
