+++ docs/source/_api_docs/config/operations.rst
@@ -1,0 +1,7 @@
+asgardpy.config.generator module
+================================
+
+.. automodule:: asgardpy.config.operations
+   :members: get_model_template, recursive_merge_dicts, deep_update
+   :undoc-members:
+   :show-inheritance:
