+++ src/asgardpy/config/generator.py
@@ -5,17 +5,22 @@ Main AsgardpyConfig Generator Module
 import json
 import logging
 import os
-from collections.abc import Mapping
 from enum import Enum
 from pathlib import Path
 
-import numpy as np
 import yaml
-from gammapy.modeling.models import CompoundSpectralModel, Models, SkyModel
+from gammapy.modeling.models import CompoundSpectralModel
 from gammapy.utils.scripts import make_path, read_yaml
 
 from asgardpy.analysis.step_base import AnalysisStepEnum
 from asgardpy.base import BaseConfig, PathType
+from asgardpy.config.operations import (
+    CONFIG_PATH,
+    check_gammapy_model,
+    compound_model_dict_converstion,
+    deep_update,
+    recursive_merge_dicts,
+)
 from asgardpy.data import (
     Dataset1DConfig,
     Dataset3DConfig,
@@ -25,16 +30,11 @@ from asgardpy.data import (
 )
 
 __all__ = [
-    "all_model_templates",
     "AsgardpyConfig",
     "GeneralConfig",
     "gammapy_model_to_asgardpy_model_config",
-    "get_model_template",
-    "recursive_merge_dicts",
     "write_asgardpy_model_to_file",
 ]
-
-CONFIG_PATH = Path(__file__).resolve().parent
 
 log = logging.getLogger(__name__)
 
@@ -69,34 +69,6 @@ class GeneralConfig(BaseConfig):
     stacked_dataset: bool = False
 
 
-def all_model_templates():
-    """
-    Collect all Template Models provided in Asgardpy, and their small tag names.
-    """
-    template_files = sorted(list(CONFIG_PATH.glob("model_templates/model_template*yaml")))
-
-    all_tags = []
-    for file in template_files:
-        all_tags.append(file.name.split("_")[-1].split(".")[0])
-    all_tags = np.array(all_tags)
-
-    return all_tags, template_files
-
-
-def get_model_template(spec_model_tag):
-    """
-    Read a particular template model yaml filename to create an AsgardpyConfig
-    object.
-    """
-    all_tags, template_files = all_model_templates()
-    new_model_file = None
-
-    for file, tag in zip(template_files, all_tags, strict=True):
-        if spec_model_tag == tag:
-            new_model_file = file
-    return new_model_file
-
-
 def check_config(config):
     """
     For a given object type, try to read it as an AsgardpyConfig object.
@@ -112,89 +84,6 @@ def check_config(config):
         raise TypeError(f"Invalid type: {config}")
 
     return AConfig
-
-
-def check_gammapy_model(gammapy_model):
-    """
-    For a given object type, try to read it as a Gammapy Models object.
-    """
-    if isinstance(gammapy_model, Models | SkyModel):
-        models_gpy = Models(gammapy_model)
-    else:
-        try:
-            models_gpy = Models.read(gammapy_model)
-        except KeyError:
-            raise TypeError("%s File cannot be read by Gammapy Models", gammapy_model) from KeyError
-
-    return models_gpy
-
-
-def recursive_merge_dicts(base_config, extra_config):
-    """
-    recursively merge two dictionaries.
-    Entries in extra_config override entries in base_config. The built-in
-    update function cannot be used for hierarchical dicts.
-
-    Also for the case when there is a list of dicts involved, one has to be
-    more careful. The extra_config may have longer list of dicts as compared
-    with the base_config, in which case, the extra items are simply added to
-    the merged final list.
-
-    Combined here are 2 options from SO.
-
-    See:
-    http://stackoverflow.com/questions/3232943/update-value-of-a-nested-dictionary-of-varying-depth/3233356#3233356
-    and also
-    https://stackoverflow.com/questions/3232943/update-value-of-a-nested-dictionary-of-varying-depth/18394648#18394648
-
-    Parameters
-    ----------
-    base_config : dict
-        dictionary to be merged
-    extra_config : dict
-        dictionary to be merged
-    Returns
-    -------
-    final_config : dict
-        merged dict
-    """
-    final_config = base_config.copy()
-
-    for key, value in extra_config.items():
-        if key in final_config and isinstance(final_config[key], list):
-            new_config = []
-
-            for key_, value_ in zip(final_config[key], value, strict=False):
-                key_ = recursive_merge_dicts(key_ or {}, value_)
-                new_config.append(key_)
-
-            # For example moving from a smaller list of model parameters to a
-            # longer list.
-            if len(final_config[key]) < len(extra_config[key]):
-                for value_ in value[len(final_config[key]) :]:
-                    new_config.append(value_)
-            final_config[key] = new_config
-
-        elif key in final_config and isinstance(final_config[key], dict):
-            final_config[key] = recursive_merge_dicts(final_config.get(key) or {}, value)
-        else:
-            final_config[key] = value
-
-    return final_config
-
-
-def deep_update(d, u):
-    """
-    Recursively update a nested dictionary.
-
-    Just like in Gammapy, taken from: https://stackoverflow.com/a/3233356/19802442
-    """
-    for k, v in u.items():
-        if isinstance(v, Mapping):
-            d[k] = deep_update(d.get(k, {}), v)
-        else:
-            d[k] = v
-    return d
 
 
 def gammapy_model_to_asgardpy_model_config(gammapy_model, asgardpy_config_file=None, recursive_merge=True):
@@ -224,22 +113,9 @@ def gammapy_model_to_asgardpy_model_config(gammapy_model, asgardpy_config_file=N
 
     # For EBL part only
     if "model1" in models_gpy_dict["components"][0]["spectral"].keys():
-        ebl_abs = models_gpy_dict["components"][0]["spectral"]["model2"]
-        ebl_abs["alpha_norm"] = ebl_abs["parameters"][0]["value"]
-        ebl_abs["redshift"] = ebl_abs["parameters"][1]["value"]
-        ebl_abs.pop("parameters", None)
-
-        models_gpy_dict["components"][0]["spectral"]["type"] = models_gpy_dict["components"][0]["spectral"][
-            "model1"
-        ]["type"]
-        models_gpy_dict["components"][0]["spectral"]["parameters"] = models_gpy_dict["components"][0]["spectral"][
-            "model1"
-        ]["parameters"]
-        models_gpy_dict["components"][0]["spectral"]["ebl_abs"] = ebl_abs
-
-        models_gpy_dict["components"][0]["spectral"].pop("model1", None)
-        models_gpy_dict["components"][0]["spectral"].pop("model2", None)
-        models_gpy_dict["components"][0]["spectral"].pop("operator", None)
+        models_gpy_dict["components"][0]["spectral"] = compound_model_dict_converstion(
+            models_gpy_dict["components"][0]["spectral"]
+        )
 
     asgardpy_config_target_dict = asgardpy_config.model_dump()["target"]
